♻️refactor: 가운데 컴포넌트 PrevGame로 분리
