+++ src/app/balance-game/[roomId]/page.tsx
@@ -1,13 +1,7 @@
 'use client';
 
-import { GameListCard, UserInfoCard, Chatting, Button } from '@/components';
-import {
-  Loader,
-  Link,
-  CheckCheck,
-  MousePointer2,
-  SlidersHorizontal,
-} from 'lucide-react';
+import { UserInfoCard, Chatting, Button } from '@/components';
+import { Link } from 'lucide-react';
 import { redirect, usePathname } from 'next/navigation';
 import { useEffect } from 'react';
 import useModalStore from '@/store/useModalStore';
@@ -18,7 +12,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import useFetchRoomDetail from '@/hooks/useFetchRoomDetail';
 import { QUERYKEY } from '@/constants/querykey';
 import { PATH } from '@/constants/router';
-import { SOCKET } from '@/constants/websocket';
+import PrevGame from './PrevGame';
 
 const WaitingRoom = () => {
   const path = usePathname();
@@ -49,42 +43,6 @@ const WaitingRoom = () => {
     }
   }, [myName]);
 
-  const isRoomManager = roomDetail?.hostName === myName;
-  const isReady = players.find((player) => player.name === myName)?.isReady;
-
-  const readyCount = players.reduce(
-    (count, { isReady }) => count + (isReady ? 1 : 0),
-    0
-  );
-  const isAllReady = readyCount === players.length;
-
-  const handleUnready = () => {
-    sendMessage({
-      destination: `${SOCKET.ENDPOINT.ROOM.UNREADY}`,
-    });
-  };
-
-  const handleReady = () => {
-    sendMessage({
-      destination: `${SOCKET.ENDPOINT.ROOM.READY}`,
-    });
-  };
-
-  const handleGameStart = () => {
-    sendMessage({
-      destination: `${SOCKET.ENDPOINT.BALANCE_GAME.START}`,
-      body: {
-        theme: 'GENERAL',
-        totalRounds: 10,
-      },
-    });
-    //@TODO: 중앙 컴포넌트 게임 화면으로 바꿔주기
-  };
-
-  const handleGameChange = () => {
-    //@TODO: 게임 변경 모달 띄워주기
-  };
-
   //@TODO: roomDetail이 없는 경우에는 스피너 컴포넌트 적용
   if (!myName || !roomDetail) {
     return;
@@ -111,80 +69,12 @@ const WaitingRoom = () => {
         ))}
       </section>
 
-      <section className="h-4/5 min-w-[45rem] max-w-[70rem] w-full flex flex-col justify-center items-center bg-container/50 rounded-lg gap-7">
-        <span className="font-semibold">잠시 후 게임이 시작됩니다.</span>
-        <GameListCard
-          title={roomDetail.game.nameKr}
-          description={roomDetail.game.descriptionKr}
-          src={roomDetail.game.thumbnailUrl}
-          className="h-16 pointer-events-none"
-        ></GameListCard>
-
-        <section className="flex flex-col gap-2">
-          {isRoomManager && isAllReady && (
-            <Button
-              className="text-base font-semibold w-[12rem]"
-              size="xl"
-              onClick={handleGameStart}
-            >
-              <div className="flex items-center justify-center gap-2">
-                <CheckCheck />{' '}
-                <span>
-                  게임 시작({readyCount}/{players.length})
-                </span>{' '}
-              </div>
-            </Button>
-          )}
-          {isRoomManager && !isAllReady && (
-            <Button
-              className="text-base font-semibold w-[12rem]"
-              size="xl"
-              variant={'waiting'}
-            >
-              <div className="flex items-center justify-center gap-2">
-                <Loader />{' '}
-                <span>
-                  준비 대기({readyCount}/{players.length})
-                </span>{' '}
-              </div>
-            </Button>
-          )}
-          {isRoomManager && (
-            <Button
-              className="text-base font-semibold w-[12rem]"
-              size="xl"
-              onClick={handleGameChange}
-            >
-              <div className="flex items-center justify-center gap-2">
-                <SlidersHorizontal /> <span>게임 변경</span>
-              </div>
-            </Button>
-          )}
-        </section>
-
-        {!isRoomManager && isReady && (
-          <Button
-            className="text-base font-semibold w-[12rem]"
-            size="xl"
-            variant={'waiting'}
-            onClick={handleUnready}
-          >
-            <div className="flex items-center justify-center gap-2">
-              <CheckCheck /> <span>준비 완료</span>
-            </div>
-          </Button>
-        )}
-        {!isRoomManager && !isReady && (
-          <Button
-            className="text-base font-semibold w-[12rem]"
-            size="xl"
-            onClick={handleReady}
-          >
-            <div className="flex items-center justify-center gap-2">
-              <MousePointer2 /> <span>준비 하기</span>
-            </div>
-          </Button>
-        )}
+      <section className="h-4/5 min-w-[45rem] max-w-[70rem] w-full bg-container/50 rounded-lg">
+        <PrevGame
+          roomDetail={roomDetail}
+          players={players}
+          sendMessage={sendMessage}
+        />
       </section>
 
       <section className="h-4/5 min-w-[15rem] max-w-[20rem]">
